自習：複数ウィンドウ操作の処理をフロント側の表示を優先した形で再実装

+++ work/app/Todo.php
@@ -25,9 +25,7 @@ class Todo
                     echo json_encode(['id' => $id]);
                     break;
                 case 'toggle':
-                    $isDone = $this->toggle();
-                    header('Content-Type: application/json');
-                    echo json_encode(['is_done' => $isDone]);
+                    $this->toggle();
                     break;
                 case 'delete':
                     $this->delete();
@@ -77,19 +75,18 @@ class Todo
             header('HTTP', true, 404);
             exit;
         }
+        // フロント側のチェック状態にDBを合わせる
+        $checkFlag = filter_input(INPUT_POST, 'checkFlag', FILTER_VALIDATE_BOOLEAN);
 
         $stmt = $this->pdo->prepare(
             "UPDATE todos
-        SET is_done = NOT is_done
+        SET is_done = :checkFlag
         WHERE id = :id"
         );
+        $stmt->bindValue('checkFlag', $checkFlag, \PDO::PARAM_BOOL);
         $stmt->bindValue('id', $id, \PDO::PARAM_INT);
         $stmt->execute();
 
-        // todoが存在した場合、該当のidの最新のis_doneの値は
-        // DB更新前に取得した$todoのis_doneを反転したものと等しい
-        // MySQLでは真偽値は1-0で管理しているためtrue-falseに変換する
-        return (boolean) !$todo->is_done;
     }
 
     private function delete()
@@ -108,6 +105,22 @@ class Todo
 
     private function purge()
     {
+        // フロント側のチェック済みidの配列
+        $appCheckedIds = array_map(
+            function ($value) {return (int) $value;},
+            explode(',', filter_input(INPUT_POST, 'appCheckedIds'))
+        );
+        // DB側のis_doneがtrueであるidの配列
+        $dbIsDoneIds = $this->pdo->query(
+            "SELECT id FROM todos WHERE is_done = 1"
+        )->fetchAll(\PDO::FETCH_COLUMN);
+        // 差異があれば404を返して処理を終了
+        $diff = array_diff($dbIsDoneIds, $appCheckedIds);
+        if (count($appCheckedIds) !== count($dbIsDoneIds) || count($diff)) {
+            header('HTTP', true, 404);
+            exit;
+        }
+        // 差異がなければDBを削除する
         $this->pdo->query("DELETE FROM todos WHERE is_done = 1");
     }
 }
